Move relation item content into sepearate component
