+++ frontend/src/components/query/query-builder/steps/relate/relate-step.tsx
@@ -15,23 +15,21 @@ import {
 import {
   Button,
   Icon,
-  IconName,
   InputGroup,
   Menu,
   MenuItem,
   Popover,
   Section,
-  Tag,
   Text,
 } from "@blueprintjs/core";
 import { ItemRenderer, Select } from "@blueprintjs/select";
 import Loading from "@/app/loading";
 import { ErrorDisplay } from "@/components/error-display";
-import { NewStepSelection } from "../query-builder";
-import SquareIcon, { SquareIconSize } from "@/components/icon/square-icon";
-import InvalidStepPopover from "../invalid-step-popover";
+import { NewStepSelection } from "../../query-builder";
+import InvalidStepPopover from "../../invalid-step-popover";
 import InferredSchemaColumnTag from "@/components/column/inferred-schema-column-tag";
 import SingleColumnSelector from "@/components/column-selectors/single-column-selector/single-column-selector";
+import RelationItemContent from "./relation-item-content";
 import { makeApiName } from "@/utils/make-friendly";
 import { useSelectedDatabase } from "@/stores";
 import { usePipelineSchema } from "@/data/use-user-query";
@@ -52,7 +50,7 @@ interface RelateStepProps {
   create?: boolean;
 }
 
-interface RelationItem {
+export interface RelationItem {
   originTable: HydratedTable;
   relatedTable: HydratedTable;
   relation: Relation;
@@ -228,11 +226,13 @@ export default function RelateStepComponent({
           <div className="flex flex-row items-center">
             <Text className="text-xl grow-0">Relate:</Text>
             <div className="flex flex-row flex-wrap items-center gap-2 mx-2 grow">
-              {renderRelationItemContent({
-                relation,
-                relatedTable,
-                originTable,
-              })}
+              <RelationItemContent
+                item={{
+                  relation,
+                  relatedTable,
+                  originTable,
+                }}
+              />
               <div className="flex flex-row items-center gap-2 flex-nowrap">
                 <Text className="flex-nowrap">on</Text>
                 <InferredSchemaColumnTag column={step!.relation.on} />
@@ -274,11 +274,13 @@ export default function RelateStepComponent({
           <div className="flex flex-row items-center">
             <Text className="text-xl grow-0">Relate:</Text>
             <div className="flex flex-row flex-wrap items-center gap-2 mx-2 grow">
-              {renderRelationItemContent({
-                relation,
-                relatedTable,
-                originTable,
-              })}
+              <RelationItemContent
+                item={{
+                  relation,
+                  relatedTable,
+                  originTable,
+                }}
+              />
               <div className="flex flex-row items-center gap-2 flex-nowrap">
                 <Text className="flex-nowrap">on</Text>
                 <InferredSchemaColumnTag column={step!.relation.on} />
@@ -402,7 +404,7 @@ export default function RelateStepComponent({
     return (
       <MenuItem
         key={`${relationItem.relation.id}.${relationItem.relatedTable.id}`}
-        text={renderRelationItemContent(relationItem)}
+        text={<RelationItemContent item={relationItem} />}
         onClick={handleClick}
         selected={_.isEqual(relationItem, selected)}
       />
@@ -432,92 +434,6 @@ export default function RelateStepComponent({
         setJoinColumn(null);
       }
     }
-  }
-
-  function getRelationIcon(item: RelationItem) {
-    switch (item.relation.type) {
-      case "one_to_one":
-        return "one-to-one";
-      case "many_to_many":
-        return "many-to-many";
-      case "one_to_many":
-        if (item.relation.table_1 === item.originTable.id) {
-          return "one-to-many";
-        } else {
-          return "many-to-one";
-        }
-    }
-  }
-
-  function renderRelationItemContent(item: RelationItem | null) {
-    if (item) {
-      const originKey =
-        item.relation.table_1 === item.originTable.id
-          ? item.relation.column_1
-          : item.relation.column_2;
-      const relatedKey =
-        item.relation.table_1 === item.relatedTable.id
-          ? item.relation.column_1
-          : item.relation.column_2;
-      return (
-        <div className="flex flex-row flex-wrap items-center gap-2">
-          <Tag
-            className="flex flex-row items-center py-1"
-            minimal
-            icon={
-              <SquareIcon
-                icon={item.relatedTable.icon as IconName}
-                color={item.relatedTable.color}
-                size={SquareIconSize.SMALL}
-              />
-            }
-          >
-            <Text className="font-bold text-md">{item.relatedTable.name}</Text>
-          </Tag>
-          <Text className="font-normal flex-nowrap">via</Text>
-          <div className="flex flex-row items-center gap-1">
-            <Tag
-              className="py-1 "
-              minimal
-              icon={
-                <SquareIcon
-                  icon={item.originTable.icon as IconName}
-                  color={item.originTable.color}
-                  size={SquareIconSize.SMALL}
-                />
-              }
-            >
-              <div className="flex flex-row items-center gap-1">
-                <Text className="font-bold text-md">
-                  {item.originTable.name}
-                </Text>
-                <Text className="font-normal text-md">{originKey}</Text>
-              </div>
-            </Tag>
-            <Icon icon={getRelationIcon(item) as IconName} color="gray" />
-            <Tag
-              className="py-1 "
-              minimal
-              icon={
-                <SquareIcon
-                  icon={item.relatedTable.icon as IconName}
-                  color={item.relatedTable.color}
-                  size={SquareIconSize.SMALL}
-                />
-              }
-            >
-              <div className="flex flex-row items-center gap-1">
-                <Text className="font-bold text-md">
-                  {item.relatedTable.name}
-                </Text>
-                <Text className="font-normal text-md">{relatedKey}</Text>
-              </div>
-            </Tag>
-          </div>
-        </div>
-      );
-    }
-    return "Select a table";
   }
 
   function selectJoinColumn(column: InferredSchemaColumn) {
@@ -590,7 +506,7 @@ export default function RelateStepComponent({
             onItemSelect={selectRelationItem}
           >
             <Button rightIcon="double-caret-vertical">
-              {renderRelationItemContent(selected)}
+              <RelationItemContent item={selected} />
             </Button>
           </Select>
           {!!selected ? (
